refactor: allow using server statically
將原本 "NeteaseCloudMusicApi/server"
掃描 modules 目錄的部分抽出 server.js，
改移到 main.js。這樣使用者就有辦法只載入
靜態的伺服器部分（不動態載入模組）。

這個 patch 的 breaking changes 不易發生。
可作為 patch version 發佈。
