+++ main.test.js
@@ -1,0 +1,18 @@
+const assert = require('assert')
+const main = require('./main')
+
+describe('methods in server.js', () => {
+  it('has serveNcmApi', () => {
+    assert.strictEqual(typeof main.serveNcmApi, 'function')
+  })
+
+  it('has getModulesDefinitions', () => {
+    assert.strictEqual(typeof main.getModulesDefinitions, 'function')
+  })
+})
+
+describe('methods in module', () => {
+  it('has activate_init_profile', () => {
+    assert.strictEqual(typeof main.activate_init_profile, 'function')
+  })
+})
